network, cache: Extract code from `infocache` to dedicated files
Move the code as is, without changing the content.
