+++ pkg/network/cache/domaininterface_test.go
